requirements.txt
streamlit 
flask 
tensorflow 
pandas 
numpy 
sounddevice 
soundfile 
pymongo 
python-socketio 
flask-socketio
librosa